move common test functions to utils
