+++ tests/downsize_test.rs
@@ -1,34 +1,17 @@
-extern crate vidalia;
 extern crate reqwest;
 extern crate multipart;
 
 use std::io::Read;
 use std::path::Path;
 
-use std::thread;
-use std::sync::{Once, ONCE_INIT};
-
 use multipart::server::Multipart as MultipartInbound;
 
-static START: Once = ONCE_INIT;
 static URL: &'static str = "http://localhost:3030";
 
-fn setup_client() -> reqwest::Client {
-    START.call_once(|| {
-        thread::spawn(vidalia::run);
-    });
-
-    reqwest::Client::new().unwrap()
-}
-
-macro_rules! fixture_path {
-    ($fixture_name:expr) => (
-        Path::new(env!("CARGO_MANIFEST_DIR"))
-            .join("tests/fixtures")
-            .join($fixture_name)
-            .to_str().unwrap()
-    )
-}
+#[macro_use]
+#[path = "utils.rs"]
+mod utils;
+use utils::{setup_client, MultipartResponse};
 
 #[test]
 fn it_downsizes_an_image() {
@@ -61,35 +44,4 @@ fn it_downsizes_an_image() {
         }).unwrap();
 
     assert_ne!(vec![] as Vec<u8>, image_buf);
-}
-
-struct MultipartResponse(reqwest::Response);
-
-impl Read for MultipartResponse {
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        self.0.read(buf)
-    }
-}
-
-impl<'r> multipart::server::HttpRequest for &'r mut MultipartResponse {
-    type Body = Self;
-
-    fn multipart_boundary(&self) -> Option<&str> {
-        const BOUNDARY: &'static str = "boundary=";
-
-        let content_type = std::str::from_utf8(
-            self.0.headers()
-                .get_raw("Content-Type").unwrap()
-                .one().unwrap())
-            .unwrap();
-
-        let boundary_start = content_type.find(BOUNDARY).unwrap() + BOUNDARY.len();
-        let boundary = &content_type[boundary_start..content_type.len()];
-
-        Some(boundary)
-    }
-
-    fn body(self) -> Self {
-        self
-    }
 }
